+++ cea/plots/sensitivity_demand_graphs.py
@@ -1,0 +1,85 @@
+# -*- coding: utf-8 -*-
+"""
+===========================
+Dynamic demand graphs algorithm
+===========================
+J. Fonseca  script development          25.08.15
+
+"""
+from __future__ import division
+
+import pandas as pd
+
+import matplotlib.pyplot as plt
+import numpy as np
+from matplotlib.collections import EllipseCollection
+from matplotlib.backends.backend_pdf import PdfPages
+
+
+__author__ = "Jimeno A. Fonseca"
+__copyright__ = "Copyright 2016, Architecture and Building Systems - ETH Zurich"
+__credits__ = ["Jimeno A. Fonseca"]
+__license__ = "MIT"
+__version__ = "0.1"
+__maintainer__ = "Daren Thomas"
+__email__ = "thomas@arch.ethz.ch"
+__status__ = "Production"
+
+
+def graph(locator, parameters):
+    for parameter in parameters:
+        pdf = PdfPages(locator.get_sensitivity_plots_file(parameter))
+        #read the mustar of morris analysis
+        data_mu = pd.read_excel(locator.get_sensitivity_output(), parameter+'mu')
+        data_sigma = pd.read_excel(locator.get_sensitivity_output(), parameter + 's')
+        var_names = data_mu.columns.values
+
+        # normalize data to maximum value
+        data_mu[var_names] = data_mu[var_names].div(data_mu[var_names].max(axis=1), axis=0)
+        data_sigma[var_names] = data_sigma[var_names].div(data_sigma[var_names].max(axis=1), axis=0)
+        # get x_names and y_names
+        # columns
+        x_names = data_mu.columns.tolist()
+        # rows
+        y_names = ['config '+str(i) for i in list(data_mu.index+1)]
+
+        # get counter (integer to create the graph)
+        x = range(len(x_names))
+        y = range(len (y_names))
+
+        X, Y = np.meshgrid(x,y)
+        XY = np.hstack((X.ravel()[:, np.newaxis], Y.ravel()[:, np.newaxis]))
+        ww = data_mu.values.tolist()
+        hh = data_sigma.values.tolist()
+        aa = X*0
+
+        fig, ax = plt.subplots()
+        ec = EllipseCollection(ww, hh, aa, units='x', offsets=XY, transOffset=ax.transData)
+        ec.set_array(np.array(ww).ravel())
+        ec.set_alpha(0.9)
+        ax.add_collection(ec)
+        ax.autoscale_view()
+        plt.xticks(np.arange(-1, max(x) + 1, 1.0))
+        plt.yticks(np.arange(-1, max(y) + 1, 1.0))
+        ax.set_xlabel('variables [-]')
+        ax.set_ylabel('configurations [-]')
+        ax.set_xticklabels([""]+x_names)
+        ax.set_yticklabels([""]+y_names)
+        cbar = plt.colorbar(ec)
+        cbar.set_label('mu_star')
+        pdf.savefig()
+        plt.close()
+        plt.clf()
+        pdf.close()
+
+def run_as_script():
+    import cea.globalvar as gv
+    import cea.inputlocator as inputlocator
+    gv = gv.GlobalVariables()
+    scenario_path = gv.scenario_reference
+    locator = inputlocator.InputLocator(scenario_path=scenario_path)
+    output_parameters = ['QHf_MWhyr', 'QCf_MWhyr', 'Ef_MWhyr', 'Ef0_kW', 'QHf0_kW', 'QCf0_kW']
+    graph(locator, output_parameters)
+
+if __name__ == '__main__':
+    run_as_script()
